requirements.txt
faker==22.0.0
kafka-python==2.0.2
pyspark==3.5.0
delta-spark==3.1.0
duckdb==0.10.0
dbt-core==1.7.0
dbt-postgres==1.7.0
great-expectations==0.18.0
apache-airflow==2.8.0
psycopg2-binary==2.9.9
pandas==2.1.4
pyarrow==14.0.1
boto3==1.34.0
pytest==7.4.4
pytest-cov==4.1.0
